+++ 2048-Arduboy/Draw.h
@@ -16,7 +16,7 @@ void EraseRect(int16_t x, int16_t y, uint8_t w, uint8_t h);
 void DrawMap(uint16_t board[DIM][DIM]);
 void Flash(uint16_t board[DIM][DIM]);
 void DrawGameState(bool running);
-void DrawScore(unsigned int score, unsigned int highscore, unsigned int biggest);
+void DrawScore(uint16_t score, uint16_t highscore, uint16_t biggest);
 uint8_t MsgBox(char* msg, uint8_t buttons);
 void DrawGameOver();
 // vim: tabstop=2:softtabstop=2:shiftwidth=2:expandtab:filetype=arduino
